chore(refactor): more intuitive organization (CPE-148)
